[refactor] Pagination Componentization
